refactor: the text editor becomes a component
178 lines of markup left App.jsx for src/ui/TextEditor.jsx. This is the
ordinary component case rather than the plumbing case: nothing in there
needs app state beyond the text being edited and the two ways of finishing
with it, so the whole boundary is five props. FONT_PRESETS, fontGroups and
TEXT_ANIM_DEFAULT it imports for itself.

Inside it, one patch() replaces twenty-four copies of

    setTextEdit(te => te ? ({ ...te, field: value }) : te)

each of which was a chance to guard the wrong thing or drop the guard. It
takes a function as well as an object, because four of the controls toggle
a field and so have to read what it was - a helper that covers most of the
cases is the kind that gets forgotten, and those four would have kept the
long form.

The typechecker caught those four: they referenced `te`, which the flat
form no longer supplies. Their update semantics are unchanged.

App.jsx 4365 -> 4192.

Verified by driving the editor rather than by the build passing: text, size
and alignment set in the panel survive a commit and reappear on reopening,
the frame draws, and bold toggles true/false/true/false - which is exactly
the path the flat-patch mistake would have broken.
