+++ src/App.jsx
@@ -7,6 +7,7 @@ import { NumField, clampNum } from './ui/NumField';
 import ColorPanel, { RECENT_SLOTS } from './ui/ColorPanel';
 import { TopBar } from './ui/TopBar';
 import { CutLayerPanel } from './ui/CutLayerPanel';
+import { TextEditor } from './ui/TextEditor';
 import { ToolsPanel } from './ui/ToolsPanel';
 import { Timeline } from './ui/Timeline';
 import { ProjectPicker, ProgressOverlay, SettingsModal, HelpModal, VideoImportModal, SceneDetectModal, LinkPromptModal, ToolKeysModal } from './ui/Modals';
@@ -3905,182 +3906,8 @@ export default function App() {
     // editing; docked next to CUT / LAYER it left the canvas a sliver. Sharing that panel's
     // space, the way the project tabs share one bar, costs the canvas nothing.
     const textEditorBody = textEdit ? (
-        <div className="text-panel-body">
-                        <textarea
-                            ref={textAreaRef}
-                            value={textEdit.text}
-                            onChange={e => setTextEdit(te => te ? ({ ...te, text: e.target.value }) : te)}
-                            onKeyDown={e => {
-                                if (e.key === 'Escape') { e.preventDefault(); e.stopPropagation(); cancelText(); }
-                                if (e.key === 'Enter' && (e.ctrlKey || e.metaKey)) { e.preventDefault(); e.stopPropagation(); commitText(); }
-                            }}
-                            placeholder={tr('텍스트 입력 (Ctrl+Enter 완료, Esc 취소)')}
-                        />
-                        <div className="text-editor-row">
-                            <div className="te-section">{tr('글꼴')}</div>
-                            <label className="text-editor-label">Size</label>
-                            {/* No range at all while editing - not per keystroke, and not
-                                on blur either. Any clamp during editing makes 100
-                                unreachable: "1" becomes 6 before the zeros arrive, and
-                                clamping when the field is left does the same thing a
-                                keystroke later if focus moves between digits. The size is
-                                held to 6..400 where it is used instead - textRender clamps
-                                for drawing, and commitText clamps what gets saved. */}
-                            <NumField
-                                value={textEdit.fontSize}
-                                onChange={v => setTextEdit(te => te ? ({ ...te, fontSize: v }) : te)}
-                                className="text-editor-num"
-                            />
-                            {(() => {
-                                const isPreset = FONT_PRESETS.some(f => f.value === textEdit.fontFamily);
-                                return (
-                                    <>
-                                        <select
-                                            className="text-editor-font"
-                                            value={isPreset ? textEdit.fontFamily : '__custom__'}
-                                            onChange={e => {
-                                                const v = e.target.value;
-                                                setTextEdit(te => te ? ({ ...te, fontFamily: v === '__custom__' ? (te.fontFamily || 'sans-serif') : v }) : te);
-                                            }}
-                                            title={tr('폰트')}
-                                        >
-                                            <option value="__custom__">Custom</option>
-                                            {/* Grouped by script: a flat list of twenty
-                                                is harder to use than a short one. Each
-                                                option previews its own face. */}
-                                            {fontGroups().map(([group, fonts]) => (
-                                                <optgroup key={group} label={group}>
-                                                    {fonts.map(f => (
-                                                        <option key={f.value} value={f.value} style={{ fontFamily: f.value }}>{f.label}</option>
-                                                    ))}
-                                                </optgroup>
-                                            ))}
-                                        </select>
-                                        {!isPreset && (
-                                            <input
-                                                className="text-editor-font"
-                                                value={textEdit.fontFamily}
-                                                onChange={e => setTextEdit(te => te ? ({ ...te, fontFamily: e.target.value }) : te)}
-                                                placeholder="Custom font-family"
-                                                title={tr('커스텀 폰트')}
-                                            />
-                                        )}
-                                    </>
-                                );
-                            })()}
-                            <input
-                                type="color"
-                                value={textEdit.color}
-                                onChange={e => setTextEdit(te => te ? ({ ...te, color: e.target.value }) : te)}
-                                className="text-editor-color"
-                                title={tr('색상')}
-                            />
-                            <button className="button" title={tr('굵게')} onClick={() => setTextEdit(te => te ? ({ ...te, bold: !te.bold }) : te)}
-                                style={{ height: 26, width: 28, padding: 0, fontWeight: 800, background: textEdit.bold ? 'hsl(var(--ui-h) var(--ui-s) 29%)' : undefined }}>B</button>
-                            <button className="button" title={tr('기울임')} onClick={() => setTextEdit(te => te ? ({ ...te, italic: !te.italic }) : te)}
-                                style={{ height: 26, width: 28, padding: 0, fontStyle: 'italic', background: textEdit.italic ? 'hsl(var(--ui-h) var(--ui-s) 29%)' : undefined }}>I</button>
-                            <select className="time-input" style={{ width: 52 }} title={tr('정렬')} value={textEdit.align || 'left'}
-                                onChange={e => setTextEdit(te => te ? ({ ...te, align: e.target.value }) : te)}>
-                                <option value="left">◧</option><option value="center">▣</option><option value="right">◨</option>
-                            </select>
-                            <select className="time-input" style={{ width: 54 }} title={tr('줄간격')} value={textEdit.lineHeight ?? 1.25}
-                                onChange={e => setTextEdit(te => te ? ({ ...te, lineHeight: +e.target.value }) : te)}>
-                                {[1, 1.15, 1.25, 1.5, 1.8, 2].map(v => <option key={v} value={v}>{v}x</option>)}
-                            </select>
-                            <select className="time-input" style={{ width: 58 }} title={tr('자간(글자 간격)')} value={textEdit.letterSpacing ?? 0}
-                                onChange={e => setTextEdit(te => te ? ({ ...te, letterSpacing: +e.target.value }) : te)}>
-                                {[-2, 0, 1, 2, 4, 8, 12].map(v => <option key={v} value={v}>{tr('자간')}{v}</option>)}
-                            </select>
-                            <div className="te-section">{tr('효과')}</div>
-                            <label className="te-check" title={tr('가독성용 외곽선')}>
-                                <input type="checkbox" checked={!!textEdit.outline} onChange={e => setTextEdit(te => te ? ({ ...te, outline: e.target.checked }) : te)} />{tr('테두리')}
-                            </label>
-                            {textEdit.outline && <input type="color" value={textEdit.outlineColor || '#ffffff'} onChange={e => setTextEdit(te => te ? ({ ...te, outlineColor: e.target.value }) : te)} className="text-editor-color" title={tr('테두리 색')} />}
-                            <label className="te-check" title={tr('그림자')}>
-                                <input type="checkbox" checked={!!textEdit.shadow} onChange={e => setTextEdit(te => te ? ({ ...te, shadow: e.target.checked }) : te)} />{tr('그림자')}
-                            </label>
-                            {textEdit.shadow && <input type="color" value={(textEdit.shadowColor || '#000000').startsWith('#') ? textEdit.shadowColor : '#000000'} onChange={e => setTextEdit(te => te ? ({ ...te, shadowColor: e.target.value }) : te)} className="text-editor-color" title={tr('그림자 색')} />}
-                            <label className="te-check" title={tr('위→아래 2색 그라데이션')}>
-                                <input type="checkbox" checked={!!textEdit.gradient} onChange={e => setTextEdit(te => te ? ({ ...te, gradient: e.target.checked }) : te)} />{tr('그라데이션')}
-                            </label>
-                            {textEdit.gradient && <input type="color" value={textEdit.color2 || '#ffffff'} onChange={e => setTextEdit(te => te ? ({ ...te, color2: e.target.value }) : te)} className="text-editor-color" title={tr('그라데이션 끝 색')} />}
-                            <label className="te-check" title={tr('글자 뒤 배경 박스')}>
-                                <input type="checkbox" checked={!!textEdit.bgColor} onChange={e => setTextEdit(te => te ? ({ ...te, bgColor: e.target.checked ? (te.bgColor || '#ffffff') : '' }) : te)} />{tr('배경')}
-                            </label>
-                            {textEdit.bgColor && <input type="color" value={textEdit.bgColor.startsWith('#') ? textEdit.bgColor : '#ffffff'} onChange={e => setTextEdit(te => te ? ({ ...te, bgColor: e.target.value }) : te)} className="text-editor-color" title={tr('배경 색')} />}
-                            <NumField className="time-input" width={54} title={tr('회전(도)')} value={textEdit.rotation ?? 0} step={5}
-                                onChange={v => setTextEdit(te => te ? ({ ...te, rotation: v }) : te)} />
-                            <NumField className="time-input" width={54} title={tr('곡률(도) — 양수는 위로 휩니다')} value={textEdit.curve ?? 0} step={10}
-                                onChange={v => setTextEdit(te => te ? ({ ...te, curve: Math.max(-180, Math.min(180, v)) }) : te)} />
-                            <label className="te-check" title={tr('좌우 반전')}>
-                                <input type="checkbox" checked={!!textEdit.flipX} onChange={e => setTextEdit(te => te ? ({ ...te, flipX: e.target.checked }) : te)} />{tr('좌우뒤집기')}
-                            </label>
-                            <label className="te-check" title={tr('상하 반전')}>
-                                <input type="checkbox" checked={!!textEdit.flipY} onChange={e => setTextEdit(te => te ? ({ ...te, flipY: e.target.checked }) : te)} />{tr('상하뒤집기')}
-                            </label>
-                            {/* Text animation, visible only during playback. */}
-                            {(() => {
-                                const an = { ...TEXT_ANIM_DEFAULT, ...(textEdit.anim || {}) };
-                                const on = !!textEdit.anim;
-                                const set = (o) => setTextEdit(te => te ? ({ ...te, anim: { ...an, ...o } }) : te);
-                                return (<>
-                                    <div className="te-section">{tr('애니메이션')}</div>
-                                    <label className="te-check" title={tr('재생할 때만 적용됩니다')}>
-                                        <input type="checkbox" checked={on}
-                                            onChange={e => setTextEdit(te => te ? ({ ...te, anim: e.target.checked ? { ...TEXT_ANIM_DEFAULT } : null }) : te)} />{tr('애니메이션')}
-                                    </label>
-                                    {on && <>
-                                        <select className="time-input" style={{ width: 74 }} title={tr('등장')} value={an.inType} onChange={e => set({ inType: e.target.value })}>
-                                            <option value="none">{tr('등장없음')}</option><option value="fade">{tr('페이드')}</option><option value="up">{tr('아래→위')}</option>
-                                            <option value="down">{tr('위→아래')}</option><option value="scale">{tr('확대')}</option><option value="blur">{tr('흐림')}</option>
-                                        </select>
-                                        <select className="time-input" style={{ width: 74 }} title={tr('퇴장')} value={an.outType} onChange={e => set({ outType: e.target.value })}>
-                                            <option value="none">{tr('퇴장없음')}</option><option value="fade">{tr('페이드')}</option><option value="up">{tr('위로')}</option>
-                                            <option value="down">{tr('아래로')}</option><option value="scale">{tr('축소')}</option><option value="blur">{tr('흐림')}</option>
-                                        </select>
-                                        {/* A stagger is a modifier: it needs an entrance or a letter effect to
-                                            spread, and typing overrides it entirely. Offering it when it can do
-                                            nothing is what made the whole feature look broken. */}
-                                        {!an.typing && (an.inType !== 'none' || an.outType !== 'none' || (an.charFx ?? 'none') !== 'none') &&
-                                        <select className="time-input" style={{ width: 92 }} title={tr('글자 하나씩 — 등장/퇴장이 글자마다 차례로 일어납니다')} value={an.charStagger ?? 0}
-                                            onChange={e => set({ charStagger: +e.target.value })}>
-                                            <option value="0">{tr('통째로')}</option><option value="0.4">{tr('글자별 약하게')}</option>
-                                            <option value="0.7">{tr('글자별')}</option><option value="0.9">{tr('글자별 크게')}</option>
-                                        </select>}
-                                        <select className="time-input" style={{ width: 74 }} title={tr('계속 반복되는 강조')} value={an.emphasis} onChange={e => set({ emphasis: e.target.value })}>
-                                            <option value="none">{tr('강조없음')}</option><option value="pulse">{tr('두근두근')}</option><option value="shake">{tr('흔들기')}</option><option value="swing">{tr('갸우뚱')}</option>
-                                        </select>
-                                        {an.emphasis !== 'none' && <>
-                                            <input type="number" className="time-input" style={{ width: 50 }} title={tr('강조 세기')} value={an.emAmount} step={5} min={0}
-                                                onChange={e => set({ emAmount: Math.max(0, +e.target.value || 0) })} />
-                                            <input type="number" className="time-input" style={{ width: 50 }} title={tr('강조 속도')} value={an.emSpeed} step={0.5} min={0}
-                                                onChange={e => set({ emSpeed: Math.max(0, +e.target.value || 0) })} />
-                                        </>}
-                                        <label className="te-check" title={tr('한 글자씩 나타남 — 등장 효과를 고르면 글자마다 그 효과로 들어옵니다')}>
-                                            <input type="checkbox" checked={!!an.typing} onChange={e => set({ typing: e.target.checked })} />{tr('타이핑')}
-                                        </label>
-                                        {an.typing && <input type="number" className="time-input" style={{ width: 56 }} title={tr('초당 글자수')} value={an.typeSpeed} step={2} min={1}
-                                            onChange={e => set({ typeSpeed: Math.max(1, +e.target.value || 1) })} />}
-                                        {/* Its own control rather than a modifier on the entrance: this is an
-                                            entrance in itself, and needing to pick a second one as well was
-                                            why ticking only 'typing' looked like it did nothing. */}
-                                        <select className="time-input" style={{ width: 96 }} title={tr('글자마다 따로 — 각 글자가 저마다 다른 곳에서 들어옵니다')} value={an.charFx ?? 'none'}
-                                            onChange={e => set({ charFx: e.target.value })}>
-                                            <option value="none">{tr('글자효과없음')}</option><option value="scatter">{tr('흩어져 모임')}</option>
-                                            <option value="drop">{tr('하나씩 떨어짐')}</option><option value="zigzag">{tr('위아래 번갈아')}</option>
-                                            <option value="spin">{tr('돌면서')}</option><option value="pop">{tr('톡톡 튀어나옴')}</option>
-                                        </select>
-                                        {(an.charFx ?? 'none') !== 'none' && <input type="number" className="time-input" style={{ width: 56 }} title={tr('글자 효과 세기 (px, 회전은 도)')} value={an.charFxAmount ?? 40} step={10} min={0}
-                                            onChange={e => set({ charFxAmount: Math.max(0, +e.target.value || 0) })} />}
-                                    </>}
-                                </>);
-                            })()}
-                            <div className="te-footer">
-                                <button className="button button-primary" onClick={commitText} style={{ height: 28, padding: '0 12px' }}>{tr('완료')}</button>
-                                <button className="button" onClick={cancelText} style={{ height: 28, padding: '0 12px' }}>{tr('취소')}</button>
-                            </div>
-                        </div>
-        </div>
+        <TextEditor textEdit={textEdit} setTextEdit={setTextEdit} textAreaRef={textAreaRef}
+            commitText={commitText} cancelText={cancelText} />
     ) : null;
 
     const cutPanelEl = (
